ДЗ 3.5 Добавлены сервисы для сохранения файлов. Добавлены DTO. Покрыл тестами что смог

+++ src/main/java/ru/hogwarts/school/services/impl/filestorage/FileSystemStorageService.java
@@ -3,45 +3,59 @@ package ru.hogwarts.school.services.impl.filestorage;
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.UrlResource;
 import org.springframework.stereotype.Service;
+import org.springframework.util.FileSystemUtils;
 import org.springframework.web.multipart.MultipartFile;
+import ru.hogwarts.school.services.api.StorageService;
+import ru.hogwarts.school.services.impl.filestorage.exception.StorageException;
 import ru.hogwarts.school.services.impl.filestorage.exception.StorageFileNotFoundException;
 
 import java.io.*;
 import java.net.MalformedURLException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.StandardOpenOption;
+import java.nio.file.*;
+import java.util.Objects;
+
+import static java.nio.file.StandardOpenOption.CREATE_NEW;
 
 @Service
-public class FileSystemStorageService {
+public class FileSystemStorageService implements StorageService {
 
-    public void save(Path filePath, MultipartFile file) throws IOException {
-        Files.createDirectories(filePath.getParent());
-        Files.deleteIfExists(filePath);
-        try (
-                InputStream is = file.getInputStream();
-                OutputStream os = Files.newOutputStream(filePath, StandardOpenOption.CREATE_NEW);
-                BufferedInputStream bis = new BufferedInputStream(is, 1024);
-                BufferedOutputStream bos = new BufferedOutputStream(os, 1024);
-        ) {
-            bis.transferTo(bos);
+    @Override
+    public void save(Path filePath, MultipartFile file) {
+        try {
+            if (file.isEmpty()) {
+                throw new StorageException("Не удалось сохранить пустой файл.");
+            }
+            Files.createDirectories(filePath.getParent());
+            Files.deleteIfExists(filePath);
+            try (
+                    InputStream is = file.getInputStream();
+                    OutputStream os = Files.newOutputStream(filePath, CREATE_NEW);
+                    BufferedInputStream bis = new BufferedInputStream(is, 1024);
+                    BufferedOutputStream bos = new BufferedOutputStream(os, 1024);
+                    ) {
+                bis.transferTo(bos);
+            }
+        } catch (IOException e) {
+            throw new StorageException("Не удалось сохранить файл.", e);
         }
     }
-
+    @Override
     public Resource loadAsResource(Path file) {
         try {
             Resource resource = new UrlResource(file.toUri());
             if (resource.exists() || resource.isReadable()) {
                 return resource;
-            }
-            else {
+            } else {
                 throw new StorageFileNotFoundException(
                         "Не удалось прочитать файл: " + file.getFileName());
             }
-        }
-        catch (MalformedURLException e) {
+        } catch (MalformedURLException e) {
             throw new StorageFileNotFoundException("Не удалось прочитать файл: " + file.getFileName(), e);
         }
+    }
+    @Override
+    public void deleteAll(Path location) {
+        FileSystemUtils.deleteRecursively(location.toFile());
     }
 
 }
